Integrate CMAC into mitosis transport.
The assumption is that AES-CMAC will be faster than HMAC-SHA256 because of the
AES hardware acceleration, however testing doesn't seem to show any difference.
AES-CMAC seems to *feel* faster, i.e. less perceived delay between button press
and glyph appearing on-screen.

During testing, it was noticed that transmit failures increased sharply as
distance increased; this may be due to electromagnetic interference in the
testing environment, or perhaps added processing delay on the receiver from
AES-CMAC vs. HMAC-SHA256.  Additional testing or perfomance diagnostics will
be needed to rule out potential delays as a factor.