Fix bug in mitosis_cmac_complete().
When reusing a mitosis_cmac_context, the plaintext_index and multiblock flag
weren't being reset in mitosis_cmac_complete(), which caused calls subsequent
to the first to generate incorrect hashes.
This commit fixes this bug by resetting that state in the complete() call.

Also added a test for this particular scenario to catch it in the future.